+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
-## 0.0.1
+## 1.0.0-beta.3
+### Fixed
+- An issue where context menus would get stuck when triggered by rundown items
+### Added
+- Time indicators in the rundown when items are scheduled or triggered
+- On end actions for playable items
+- An onEnd event
+- A note regarding supported Caspar server versions in settings
+
+## 1.0.0-beta.2
+### Changed
+- Updated UI
+### Fixed
+- General bug fixes
+
+## 1.0.0-beta.1
 ### Changed
 - Initial development version